+++ imagebutler/__init__.py
@@ -5,6 +5,6 @@ from . import models
 from . import apis, commands, views
 
 
-__version__ = '0.0.2'
+__version__ = '0.0.3'
 
 app.register_blueprint(api_bp)