+++ imagebutler/job_workers.py
@@ -2,4 +2,11 @@
 
 
 def worker_do_cache_redis(caching_object):
+    """Just a wrapper function to cache the caching object. It's the way
+    RQ act."""
     return caching_object
+
+
+def worker_undo_cached_redis(caching_object):
+    """Just a wrapper function to clean RQ's jobs' results."""
+    return caching_object.delete()