+++ HISTORY.rst
@@ -1,6 +1,15 @@
 History
 =======
 
+0.0.3 (Dec 10th, 2017)
+----------------------
+
+- Integrating with tox and travis
+- Add unit tests for imagebutler.utils module
+- API for add, modify, delete image
+- API for listing all the user's images
+- Add thumbnail column to Image table
+
 0.0.2 (Nov 13th, 2017)
 ----------------------